[account] move seed tests to seed_test.go
